reductive changes

+++ src/V3OrderMoveGraph.h
@@ -54,8 +54,8 @@ class OrderMoveDomScope final {
 
         struct Hash final {
             size_t operator()(const DomScopeMapKey& key) const {
-                V3Hash hash{reinterpret_cast<uintptr_t>(key.m_domainp)};
-                hash += reinterpret_cast<uintptr_t>(key.m_scopep);
+                V3Hash hash{static_cast<uint32_t>(reinterpret_cast<uintptr_t>(key.m_domainp))};
+                hash += static_cast<uint32_t>(reinterpret_cast<uintptr_t>(key.m_scopep));
                 return hash.value();
             }
         };
